Add router for /ids to paginate to previous days.
