+++ src/views/home.ejs
@@ -18,7 +18,7 @@
       <h1>Daily Prompt</h1>
 
       <h2 id="todaysPrompt">
-        Write whatever you want.
+        <%= prompt %>
       </h2>
     </header>
